split up source file + issolate GLSL emit path change.
